Clean up MME usages
