+++ Sources/MapboxCoreNavigation/EventDetails.swift
@@ -4,7 +4,6 @@ import Polyline
 import UIKit
 import AVFoundation
 import MapboxDirections
-import MapboxMobileEvents
 
 protocol EventDetails: Encodable {
     var event: String? { get set }
@@ -174,5 +173,19 @@ extension AVAudioSession {
             return "speaker"
         }
         return "unknown"
+    }
+}
+
+extension UIDevice {
+    @nonobjc var machine: String {
+        get {
+            var systemInfo = utsname()
+            uname(&systemInfo)
+            let machineMirror = Mirror(reflecting: systemInfo.machine)
+            return machineMirror.children.reduce("") { (identifier: String, element: Mirror.Child) in
+                guard let value = element.value as? Int8, value != 0 else { return identifier }
+                return identifier + String(UnicodeScalar(UInt8(value)))
+            }
+        }
     }
 }
